+++ noodles-sam/src/async/writer.rs
@@ -1,0 +1,109 @@
+use tokio::io::{self, AsyncWrite, AsyncWriteExt};
+
+use crate::{Header, Record};
+
+/// An async SAM writer.
+pub struct Writer<W>
+where
+    W: AsyncWrite,
+{
+    inner: W,
+}
+
+impl<W> Writer<W>
+where
+    W: AsyncWrite + Unpin,
+{
+    /// Creates an async SAM writer.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use noodles_sam as sam;
+    /// let writer = sam::AsyncWriter::new(Vec::new());
+    /// ```
+    pub fn new(inner: W) -> Self {
+        Self { inner }
+    }
+
+    /// Returns a reference to the underlying writer.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use noodles_sam as sam;
+    /// let writer = sam::AsyncWriter::new(Vec::new());
+    /// assert!(writer.get_ref().is_empty());
+    /// ```
+    pub fn get_ref(&self) -> &W {
+        &self.inner
+    }
+
+    /// Returns the underlying writer.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use noodles_sam as sam;
+    /// let writer = sam::AsyncWriter::new(Vec::new());
+    /// assert!(writer.into_inner().is_empty());
+    /// ```
+    pub fn into_inner(self) -> W {
+        self.inner
+    }
+
+    /// Writes a SAM header.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # use std::io;
+    /// #
+    /// # #[tokio::main]
+    /// # async fn main() -> io::Result<()> {
+    /// use noodles_sam as sam;
+    ///
+    /// let mut writer = sam::AsyncWriter::new(Vec::new());
+    ///
+    /// let header = sam::Header::builder().add_comment("noodles-sam").build();
+    /// writer.write_header(&header).await?;
+    ///
+    /// assert_eq!(writer.get_ref(), b"@CO\tnoodles-sam\n");
+    /// # Ok(())
+    /// # }
+    /// ```
+    pub async fn write_header(&mut self, header: &Header) -> io::Result<()> {
+        let raw_header = header.to_string();
+        self.inner.write_all(raw_header.as_bytes()).await
+    }
+
+    /// Writes a SAM record.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # use std::io;
+    /// #
+    /// # #[tokio::main]
+    /// # async fn main() -> io::Result<()> {
+    /// use noodles_sam as sam;
+    ///
+    /// let mut writer = sam::AsyncWriter::new(Vec::new());
+    ///
+    /// let record = sam::Record::default();
+    /// writer.write_record(&record).await?;
+    ///
+    /// assert_eq!(writer.get_ref(), b"*\t4\t*\t0\t255\t*\t*\t0\t0\t*\t*\n");
+    /// # Ok(())
+    /// # }
+    /// ```
+    pub async fn write_record(&mut self, record: &Record) -> io::Result<()> {
+        const LINE_FEED: u8 = b'\n';
+
+        let raw_record = record.to_string();
+        self.inner.write_all(raw_record.as_bytes()).await?;
+        self.inner.write_all(&[LINE_FEED]).await?;
+
+        Ok(())
+    }
+}