+++ noodles-sam/CHANGELOG.md
@@ -14,7 +14,7 @@
 
     This previously panicked if the ID was not set.
 
-  * sam/header/read_group/builder: Return error from `build`.
+  * header/read_group/builder: Return error from `build`.
 
     This previously panicked if the ID was not set.